# docs/swagger/swagger.json
{
    "schemes": [
        "http"
    ],
    "swagger": "2.0",
    "info": {
        "description": "Система управления мероприятиями: события, площадки, билеты, отчёты, аудит.",
        "title": "Time2Meet API",
        "contact": {},
        "version": "0.1.0"
    },
    "host": "localhost:8080",
    "basePath": "/api/v1",
    "paths": {
...
        "/batch/import/events": {
            "post": {
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "batch"
                ],
                "summary": "Батч-импорт мероприятий",
                "parameters": [
                    {
                        "type": "string",
                        "description": "User ID (UUID) for audit",
                        "name": "X-User-Id",
                        "in": "header"
                    },
                    {
                        "description": "Пакет мероприятий",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/handler.importEventsRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/batchimport.Result"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/handler.ErrorResponse"
                        }
                    },
                    "409": {
                        "description": "Conflict",
                        "schema": {
                            "$ref": "#/definitions/handler.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/handler.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/batch/import/tickets": {
            "post": {
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "batch"
                ],
                "summary": "Батч-импорт билетов",
                "parameters": [
                    {
                        "type": "string",
                        "description": "User ID (UUID) for audit",
                        "name": "X-User-Id",
                        "in": "header"
                    },
                    {
                        "description": "Пакет билетов",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/handler.importTicketsRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/batchimport.Result"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/handler.ErrorResponse"
                        }
                    },
                    "409": {
                        "description": "Conflict",
                        "schema": {
                            "$ref": "#/definitions/handler.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/handler.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/batch/import/users": {
            "post": {
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "batch"
                ],
                "summary": "Батч-импорт пользователей",
                "parameters": [
                    {
                        "type": "string",
                        "description": "User ID (UUID) for audit",
                        "name": "X-User-Id",
                        "in": "header"
                    },
                    {
                        "description": "Пакет пользователей",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/handler.importUsersRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/batchimport.Result"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/handler.ErrorResponse"
                        }
                    },
                    "409": {
                        "description": "Conflict",
                        "schema": {
                            "$ref": "#/definitions/handler.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/handler.ErrorResponse"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
        "apperror.Code": {
            "type": "string",
            "enum": [
                "not_found",
                "conflict",
                "validation",
                "unauthorized",
                "forbidden",
                "internal",
                "invalid_state",
                "unavailable"
            ],
            "x-enum-varnames": [
                "CodeNotFound",
                "CodeConflict",
                "CodeValidation",
                "CodeUnauthorized",
                "CodeForbidden",
                "CodeInternal",
                "CodeInvalidState",
                "CodeUnavailable"
            ]
        },
        "batchimport.BatchError": {
            "type": "object",
            "properties": {
                "error": {
                    "type": "string"
                },
                "index": {
                    "type": "integer"
                }
            }
        },
        "batchimport.ImportEventsItem": {
            "type": "object",
            "properties": {
                "cover_image": {
                    "type": "string"
                },
                "description": {
                    "type": "string"
                },
                "is_public": {
                    "type": "boolean"
                },
                "max_participants": {
                    "type": "integer"
                },
                "organizer_id": {
                    "type": "string"
                },
                "status": {
                    "type": "string"
                },
                "title": {
                    "type": "string"
                }
            }
        },
        "batchimport.ImportTicketsItem": {
            "type": "object",
            "properties": {
                "amount_paid": {
                    "type": "string"
                },
                "buyer_id": {
                    "type": "string"
                },
                "purchase_date": {
                    "type": "string"
                },
                "qr_code": {
                    "type": "string"
                },
                "status": {
                    "type": "string"
                },
                "ticket_type_id": {
                    "type": "string"
                }
            }
        },
        "batchimport.ImportUsersItem": {
            "type": "object",
            "properties": {
                "email": {
                    "type": "string"
                },
                "full_name": {
                    "type": "string"
                },
                "password_hash": {
                    "type": "string"
                },
                "phone": {
                    "type": "string"
                },
                "role": {
                    "type": "string"
                }
            }
        },
        "batchimport.Result": {
            "type": "object",
            "properties": {
                "errors": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/batchimport.BatchError"
                    }
                },
                "failed": {
                    "type": "integer"
                },
                "success": {
                    "type": "integer"
                },
                "total": {
                    "type": "integer"
                }
            }
        },
...
        "handler.ErrorResponse": {
            "type": "object",
            "properties": {
                "code": {
                    "$ref": "#/definitions/apperror.Code"
                },
                "message": {
                    "type": "string"
                }
            }
        },
...
        "handler.importEventsRequest": {
            "type": "object",
            "required": [
                "items"
            ],
            "properties": {
                "continue_on_error": {
                    "type": "boolean"
                },
                "items": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/batchimport.ImportEventsItem"
                    }
                }
            }
        },
        "handler.importTicketsRequest": {
            "type": "object",
            "required": [
                "items"
            ],
            "properties": {
                "continue_on_error": {
                    "type": "boolean"
                },
                "items": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/batchimport.ImportTicketsItem"
                    }
                }
            }
        },
        "handler.importUsersRequest": {
            "type": "object",
            "required": [
                "items"
            ],
            "properties": {
                "continue_on_error": {
                    "type": "boolean"
                },
                "items": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/batchimport.ImportUsersItem"
                    }
                }
            }
        },
...
    }
}